mysql-connector/requirements.txt
fastapi==0.63.0
mysql-connector-python==8.0.22
protobuf==3.14.0
pydantic==1.7.3
six==1.15.0
starlette==0.13.6
uvicorn==0.11.7
